--- src/config.rs
@@ -1,9 +1,0 @@
-mod arg;
-mod pacman;
-mod file;
-
-pub(crate) use arg::Arg;
-pub(crate) use pacman::Config as PacmanConfig;
-pub(crate) use file::Config;
-pub(crate) use file::DepHashStrategy;
-pub(crate) use file::Pkgbuild;
